rename cli param files
